Prevent caching of prebind requests
If the connection fails from the prebind session, when it tries to start another BOSH session, the response from the prebind url gets cached, resulting in the old BOSH session info being returned. So any subsequent reconnect attempts failed. This happened to me in IE and adding the cache flag on the ajax call fixed the issue.